+++ src/core/interaction/pickingobject.cpp
@@ -23,7 +23,7 @@ const DataVec3UINT8::type& PickingObject::getPickingColorAsUINT8() const{
     return colorUINT8_; 
 }
 
-const ivec2& PickingObject::getPickingMove() const{
+const vec2& PickingObject::getPickingMove() const{
     return move_;
 }
 
@@ -31,7 +31,7 @@ void PickingObject::picked() const{
     onPickedCallback_->invoke();
 }
 
-void PickingObject::setPickingMove(ivec2 move){
+void PickingObject::setPickingMove(vec2 move){
     move_ = move;
 }
